requirements.txt
selenium
http
pathlib
SQLAlchemy==0.9.9
passlib==1.7.1
zope.sqlalchemy<=1.0.0
